[REFACTOR]: 파일구조 리펙토링
