+++ lib/provider.dart
@@ -1,4 +1,4 @@
-import 'package:alpha_fe/pages/add_page/add_page_0/add_page_0_view_model/add_page_0_view_model.dart';
+import 'package:alpha_fe/pages/add_page/add_page_0/view_model/add_page_0_view_model.dart';
 import 'package:alpha_fe/pages/home_page/home_page_view_model/plan_view_model.dart';
 import 'package:alpha_fe/pages/home_page/home_page_view_model/recommend_place_view_model.dart';
 import 'package:alpha_fe/pages/home_page/home_page_view_model/save_tour_course_view_model.dart';
